AG-16076: combine idkey examples with base variants using matching mode selector
