--- packages/ag-charts-website/src/content/docs/high-frequency-data-test/_examples/high-freq-line-idkey/index.html
@@ -1,20 +1,0 @@
-<div class="example-controls">
-    <div class="controls-row">
-        <button id="toggleFeedBtn" onclick="toggleFeed()">Start Feed</button>
-        <button id="toggleRapidFeedBtn" onclick="toggleRapidFeed()">Start Rapid Feed</button>
-        <button onclick="tickFeed()">Tick Feed</button>
-        <select onchange="use(this.value)">
-            <option value="applyTransaction-append">applyTransaction (append)</option>
-            <option value="applyTransaction-prepend">applyTransaction (prepend)</option>
-            <option value="applyTransaction-remove-first">applyTransaction (remove first by ID)</option>
-            <option value="applyTransaction-remove-last">applyTransaction (remove last by ID)</option>
-            <option value="updateDelta-append">updateDelta (append)</option>
-            <option value="updateDelta-prepend">updateDelta (prepend)</option>
-            <option value="updateDelta-remove-first">updateDelta (remove first)</option>
-            <option value="updateDelta-remove-last">updateDelta (remove last)</option>
-        </select>
-        <span id="cpuUsage" style="margin-left: 10px">CPU: 0%</span>
-        <span style="margin-left: 10px; color: #666; font-style: italic">dataIdKey mode</span>
-    </div>
-</div>
-<div id="myChart"></div>
