--- packages/ag-charts-website/src/content/docs/high-frequency-data-test/_examples/high-freq-random-update-idkey/index.html
@@ -1,46 +1,0 @@
-<div class="example-controls">
-    <div class="controls-row">
-        <button id="toggleBtn" onclick="toggleUpdates()">Start Updates</button>
-        <button id="tickBtn" onclick="tickUpdate()">Tick</button>
-        <select id="updateCountSelect" onchange="updateUpdateCount(this.value)">
-            <option value="10">10 updates</option>
-            <option value="100" selected>100 updates</option>
-            <option value="500">500 updates</option>
-            <option value="1000">1000 updates</option>
-            <option value="5000">5000 updates</option>
-        </select>
-        <select id="methodSelect" onchange="updateMethod(this.value)">
-            <option value="applyTransaction">applyTransaction()</option>
-            <option value="updateDelta">updateDelta()</option>
-        </select>
-        <select id="frequencySelect" onchange="updateFrequency(this.value)">
-            <option value="raf">requestAnimationFrame()</option>
-            <option value="50">50ms</option>
-            <option value="100">100ms</option>
-            <option value="200" selected>200ms</option>
-            <option value="500">500ms</option>
-            <option value="1000">1000ms</option>
-        </select>
-        <select id="seriesTypeSelect" onchange="updateSeriesType(this.value)">
-            <option value="line" selected>Line</option>
-            <option value="area">Area</option>
-            <option value="bar">Bar</option>
-            <option value="bubble">Bubble</option>
-            <option value="ohlc">OHLC</option>
-            <option value="candlestick">Candlestick</option>
-            <option value="range-bar">Range Bar</option>
-            <option value="range-area">Range Area</option>
-        </select>
-        <select id="axisTypeSelect" onchange="updateAxisType(this.value)">
-            <option value="time" selected>time</option>
-            <option value="ordinal-time">ordinal-time</option>
-            <option value="unit-time">unit-time</option>
-        </select>
-        <span id="cpuUsage" style="margin-left: 10px">CPU: 0%</span>
-        <span id="fpsCounter" style="margin-left: 10px">FPS: 0</span>
-        <span id="dataCount" style="margin-left: 10px">Points: 100000</span>
-        <span id="updateCountDisplay" style="margin-left: 10px">Updates: 100</span>
-        <span style="margin-left: 10px; color: #666; font-style: italic">dataIdKey mode</span>
-    </div>
-</div>
-<div id="myChart"></div>
