+++ packages/ag-charts-website/src/content/docs/high-frequency-data-test/_examples/high-freq-random-update/index.html
@@ -1,7 +1,6 @@
 <div class="example-controls">
     <div class="controls-row">
         <button id="toggleBtn" onclick="toggleUpdates()">Start Updates</button>
-        <button id="tickBtn" onclick="tickUpdate()">Tick</button>
         <select id="updateCountSelect" onchange="updateUpdateCount(this.value)">
             <option value="10">10 updates</option>
             <option value="100" selected>100 updates</option>
@@ -35,6 +34,10 @@
             <option value="time" selected>time</option>
             <option value="ordinal-time">ordinal-time</option>
             <option value="unit-time">unit-time</option>
+        </select>
+        <select id="matchingModeSelect" onchange="setMatchingMode(this.value)">
+            <option value="ref" selected>Matching: ref</option>
+            <option value="id">Matching: dataIdKey</option>
         </select>
         <span id="cpuUsage" style="margin-left: 10px">CPU: 0%</span>
         <span id="fpsCounter" style="margin-left: 10px">FPS: 0</span>
