+++ packages/ag-charts-website/src/content/docs/high-frequency-data-test/_examples/high-freq-random-update/main.ts
@@ -3,11 +3,13 @@ import { type AgCartesianChartOptions, type AgCartesianSeriesOptions, AgCharts }
 (window as any).agChartsDebug = ['scene:stats'];
 
 type ValueDatum = {
+    id: number;
     timestamp: number;
     value: number;
 };
 
 type OhlcDatum = {
+    id: number;
     timestamp: number;
     open: number;
     high: number;
@@ -16,12 +18,14 @@ type OhlcDatum = {
 };
 
 type RangeDatum = {
+    id: number;
     timestamp: number;
     low: number;
     high: number;
 };
 
 type BubbleDatum = {
+    id: number;
     timestamp: number;
     value: number;
     size: number;
@@ -142,12 +146,15 @@ function setConfigValue<K extends keyof FormConfig>(key: K, value: FormConfig[K]
 let currentSeriesType: SeriesType;
 let currentAxisType: AxisType;
 
+let matchingMode: 'ref' | 'id' = 'ref';
+
 function generateValueDatum(index: number): ValueDatum {
     const timestamp = START_TIMESTAMP + index * DATA_INTERVAL_MS;
     const trend = Math.sin(index / 240) * 40 + Math.cos(index / 80) * 25;
     const volatility = Math.sin(index / 15) * 5;
     const baseline = 1_000 + index * 0.02;
     return {
+        id: index,
         timestamp,
         value: Number((baseline + trend + volatility).toFixed(2)),
     };
@@ -172,6 +179,7 @@ function generateOhlcDatum(index: number, previousClose?: number): { datum: Ohlc
 
     return {
         datum: {
+            id: index,
             timestamp,
             open: Number(open.toFixed(2)),
             high,
@@ -190,6 +198,7 @@ function generateRangeDatum(index: number): RangeDatum {
     const midValue = baseline + trend + volatility;
     const range = 10 + Math.abs(Math.sin(index / 30)) * 20;
     return {
+        id: index,
         timestamp,
         low: Number((midValue - range / 2).toFixed(2)),
         high: Number((midValue + range / 2).toFixed(2)),
@@ -202,6 +211,7 @@ function generateBubbleDatum(index: number): BubbleDatum {
     const volatility = Math.sin(index / 15) * 5;
     const baseline = 1_000 + index * 0.02;
     return {
+        id: index,
         timestamp,
         value: Number((baseline + trend + volatility).toFixed(2)),
         size: 5 + Math.abs(Math.sin(index / 50)) * 20,
@@ -276,6 +286,64 @@ function mutateDatum(item: Datum) {
         mutateBubbleDatum(item as BubbleDatum);
     } else {
         mutateValueDatum(item as ValueDatum);
+    }
+}
+
+// ID-based replacement functions — create new objects instead of mutating in place
+function replaceValueDatum(item: ValueDatum): ValueDatum {
+    const change = (Math.random() - 0.5) * 200;
+    return {
+        id: item.id,
+        timestamp: item.timestamp,
+        value: Number((item.value + change).toFixed(2)),
+    };
+}
+
+function replaceOhlcDatum(item: OhlcDatum): OhlcDatum {
+    const change = (Math.random() - 0.5) * 100;
+    const open = Number((item.open + change).toFixed(2));
+    const close = Number((item.close + change).toFixed(2));
+    return {
+        id: item.id,
+        timestamp: item.timestamp,
+        open,
+        close,
+        high: Number((Math.max(open, close) + Math.random() * 20).toFixed(2)),
+        low: Number((Math.min(open, close) - Math.random() * 20).toFixed(2)),
+    };
+}
+
+function replaceRangeDatum(item: RangeDatum): RangeDatum {
+    const change = (Math.random() - 0.5) * 200;
+    let low = Number((item.low + change).toFixed(2));
+    let high = Number((item.high + change).toFixed(2));
+    if (low > high) {
+        const temp = low;
+        low = high;
+        high = temp;
+    }
+    return { id: item.id, timestamp: item.timestamp, low, high };
+}
+
+function replaceBubbleDatum(item: BubbleDatum): BubbleDatum {
+    const change = (Math.random() - 0.5) * 200;
+    return {
+        id: item.id,
+        timestamp: item.timestamp,
+        value: Number((item.value + change).toFixed(2)),
+        size: Math.max(1, item.size + (Math.random() - 0.5) * 10),
+    };
+}
+
+function replaceDatum(item: Datum): Datum {
+    if (currentSeriesType === 'ohlc' || currentSeriesType === 'candlestick') {
+        return replaceOhlcDatum(item as OhlcDatum);
+    } else if (currentSeriesType === 'range-bar' || currentSeriesType === 'range-area') {
+        return replaceRangeDatum(item as RangeDatum);
+    } else if (currentSeriesType === 'bubble') {
+        return replaceBubbleDatum(item as BubbleDatum);
+    } else {
+        return replaceValueDatum(item as ValueDatum);
     }
 }
 
@@ -536,12 +604,21 @@ async function updateRandomSubset() {
         indicesToUpdate.add(Math.floor(Math.random() * data.length));
     }
 
-    // Mutate and collect items to update
     const itemsToUpdate: Datum[] = [];
-    for (const idx of indicesToUpdate) {
-        const item = data[idx];
-        mutateDatum(item);
-        itemsToUpdate.push(item);
+    if (matchingMode === 'id') {
+        // ID-based: create replacement objects instead of mutating in place
+        for (const idx of indicesToUpdate) {
+            const replacement = replaceDatum(data[idx]);
+            data[idx] = replacement;
+            itemsToUpdate.push(replacement);
+        }
+    } else {
+        // Ref-based: mutate existing objects in place
+        for (const idx of indicesToUpdate) {
+            const item = data[idx];
+            mutateDatum(item);
+            itemsToUpdate.push(item);
+        }
     }
 
     await dispatchUpdate(itemsToUpdate);
@@ -705,6 +782,22 @@ async function setSeriesType(newSeriesType: SeriesType) {
     }
 }
 
+function setMatchingMode(mode: string) {
+    const wasRunning = isRunning;
+    if (wasRunning) stopUpdates();
+
+    matchingMode = mode as 'ref' | 'id';
+
+    // Recreate chart data with or without dataIdKey
+    const seedResult = createSeedData(data.length);
+    data = seedResult.data;
+    chart.update({ data, dataIdKey: matchingMode === 'id' ? 'id' : undefined });
+
+    resetCpuIndicator();
+    resetFpsCounter();
+    if (wasRunning) startUpdates();
+}
+
 let axisTypeUpdateInProgress = false;
 
 async function setAxisType(newAxisType: AxisType) {
@@ -829,6 +922,9 @@ window.addEventListener('hashchange', () => {
 };
 (window as any).tickUpdate = () => {
     void updateRandomSubset();
+};
+(window as any).setMatchingMode = (mode: string) => {
+    setMatchingMode(mode);
 };
 
 export {};
